+++ Src/DasherCore/ButtonMultiPress.cpp
@@ -46,9 +46,9 @@ void CButtonMultiPress::KeyDown(unsigned long iTime, int iId, CDasherView *pView
     else
     {
       //previous presses should not be treated as such....
-      RevertPresses(m_deQueueTimes.size());
+      RevertPresses(static_cast<int>(m_deQueueTimes.size()));
       //...but should be combined with this one into a new event (type = #presses)
-      ButtonEvent(iTime, iId, m_deQueueTimes.size()+1, pModel);
+      ButtonEvent(iTime, iId, static_cast<int>(m_deQueueTimes.size())+1, pModel);
       if (m_deQueueTimes.size() >= maxClickCount() - 1)
 	m_deQueueTimes.clear(); //final press
       else //may still be more presses to come
